agregando Constructor inventarios

+++ Parcial1-LeonardoEmil/Entidades/Productos.cs
@@ -13,8 +13,8 @@ namespace Parcial1_LeonardoEmil.Entidades
         public int ProductoId { get; set; }
         public string Descripcion { get; set; }
         public int Existencia { get; set; }
-        public int Costo { get; set; }
-        public int ValorInventario { get; set; }
+        public float Costo { get; set; }
+        public float ValorInventario { get; set; }
 
         public Productos()
         {
